Assignment 5/requirements.txt
azure-identity
azure-mgmt-compute
azure-mgmt-monitor
azure-mgmt-resource
tabulate
pytz
